remove local.best, cache entire tree branch in local, reduce tickrange

+++ test/core/TakerOperations.t.sol
@@ -519,7 +519,8 @@ contract TakerOperationsTest is MangroveTest {
     assertTrue(successes == 0, "order should fail");
   }
 
-  function test_snipe_correct_amount_auto(uint96 makerWants, uint96 makerGives, uint96 factor1, uint16 pc) public {
+  // FIXME restricting to uint72 so maximum price is not reached
+  function test_snipe_correct_amount_auto(uint72 makerWants, uint72 makerGives, uint72 factor1, uint16 pc) public {
     vm.assume(factor1 > 0);
     vm.assume(makerWants > 0);
     vm.assume(makerGives > 0);
